% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremi.R
\name{analyze.2order}
\alias{analyze.2order}
\title{DOREMI second order analysis function}
\usage{
analyze.2order(
  data,
  id = NULL,
  input = NULL,
  time = NULL,
  signal,
  dermethod = "gold",
  derparam = 3,
  order = 2,
  verbose = FALSE
)
}
\arguments{
\item{data}{Is a data frame containing at least one column, that is the signal to be analyzed.}

\item{id}{Is a STRING containing the NAME of the column of data containing the identifier of the individual.
If this parameter is not entered when calling the function, a single individual is assumed and a linear regression is done instead
of the linear mixed-effects regression.}

\item{input}{Is a STRING or a VECTOR OF STRINGS containing the NAME(s) of data column(s) containing the EXCITATION vector(s).
If this parameter is not entered when calling the function,
the excitation is assumed to be unknown. In this case, the linear mixed-effect regression is still carried out but no coefficient is calculated
for the excitation term. If no excitation term is supplied, one of the initial conditions is different from 0 (signal or derivative) and xi<1 the function will estimate
a damped linear oscillator (DLO)}

\item{time}{Is a STRING containing the NAME of the column of data containing the time vector. If this parameter is not entered when calling the function,
it is assumed that time steps are of 1 unit and the time vector is generated internally in the function.}

\item{signal}{Is a STRING containing the NAME of the column of the data frame containing the SIGNAL to be studied.}

\item{dermethod}{is the derivative estimation method. The following methods are available: "gold","glla" and "fda"}

\item{derparam}{If dermethod "glla" or "gold" are chosen, it is the embedding number, a positive integer containing the number of points to be used for the calculation of the derivatives.
Its value by default is 3 as at least three points are needed for the calculation of the second derivative. If dermethod "fda" is chosen, this parameter is
spar, the parameter related to the smoothing parameter lambda used in the penalization function of the function \code{smooth.spline} to estimate the derivative via splines (Functional Data Analysis)}

\item{order}{is the maximum order of the derivative to estimate. Using an order higher than that of the maximum derivative to estimate (1 in first order differential equations and
2 in second order differential equations), for instance, order=4 might enhance derivative estimation (see \href{https://doi.org/10.1080/00273171.2015.1123138}{Chow et al.(2016)})}

\item{verbose}{Is a boolean that displays status messages of the function when set to 1.}
}
\value{
Returns a summary of the fixed components for the coefficients: damping factor, period, gain and equilibrium value and the R2 resulting from the estimation.
The estimation is performed using the function lmer if there are several individuals or lm if there is only one.
With the above estimated parameters, the estimated signal is reconstructed for
each individual by using the generate.2order function of this package (based on deSolve's ode function).
The function returns five objects:
\enumerate{
 \item data- A data.frame including the input data, the intermediate calculations used to prepare the variables for
  the fit and the estimated trajectories for each individual.

    signal_rollmean - calculation of the moving average of the signal over embedding points.

    signal_derivate1 - calculation of the first derivative of the signal with the chosen method in embedding points/with smoothing parameter spar

    signal_derivate2 - calculation of the second derivative of the signal with the chosen method in embedding points/with smoothing parameter spar

    time_derivate - calculation of the moving average of the time vector over embedding points.

    input_rollmean - calculation of the moving average of the excitation vector over embedding points.

    estimated- the estimated signal calculated using deSolve's ode function with a second order model, the excitation provided as input and the
    coefficients obtained from the fit.
 \item resultid- A data.frame including for each individual, listed by id number, the coefficients calculated (thus fixed + random component)
 \item resultmean- A data.frame including the fixed effects of the coefficients mentioned above and the R2 resulting from the estimation
 \item regression- A list containing the summary of the linear mixed-effects regression.

 As seen in the Description section, the print method by default prints only the resultmean element. Each one of the other objects
 can be accessed by indicating $ and their name after the result, for instance, for a DOREMI object called "result", it is possible
 to access the regression summary by typing result$regression.
 \item embedding - contains the embedding number used to generate the results (same as function input argument). Will only appear in the output if the
 derivative estimation method chosen is "glla" or "gold".
 \item spar -
}
}
\description{
\code{analyze.2order}  estimates the coefficients of a second order differential equation of the form:
\deqn{\frac{d^2y}{dt} + 2\zeta\omega_{n}\frac{dy}{dt} + \omega_{n}^2 (y - y_{eq}) = k*u(t) }
Where y(t) is the individual's signal, \eqn{\dot{y}(t)} is the derivative and u(t) is the excitation.
The function estimates the coefficients \zeta, \omega_{n}, k and y_{eq} using a two step procedure in
which the user can choose the derivative estimation method (through the parameter dermethod) and then the
coefficients are then estimated through a linear mixed-effect model.
}
\examples{
myresult <- analyze.2order(data = cardio,
                  id = "id",
                  input = "load",
                  time = "time",
                  signal = "hr",
                  dermethod = "gold",
                  derparam = 5)
}
\seealso{
\code{\link{calculate.gold}\link{calculate.glla}\link{calculate.fda}} to compute the derivatives, for details on embedding.
}
\keyword{analysis,}
\keyword{order}
\keyword{second}
