% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremi.R
\name{calculate_gold}
\alias{calculate_gold}
\title{Calculation of derivatives using the Gold method}
\usage{
calculate_gold(signal, time, embedding = 2)
}
\arguments{
\item{signal}{is a vector containing the data from which the derivative is estimated.}

\item{time}{is a vector containing the time values corresponding to the data contained in signal. Thus, they both have the same length.}

\item{embedding}{is a signed integer indicating the number of points to consider for derivative calculation. Embedding must be greater than 2 because at least
two points are needed for the calculation of the first derivative and at least 3 for the calulation of the second derivative. If it is not the case, the
function displays an error message. If not specified, embedding=2 by default.}
}
\value{
Returns a list containing three columns:
dtime- contains the time values in which the derivative was calculated. That is, the moving average of the input time over embedding points.

dsignal- is a data.frame containing three columns and the same number of rows as the signal.
The first column is the moving average of the signal over embedding points, the second is the first derivative,
and the third is the second derivative.

embedding- contains the number of points used for the derivative calculation, which is constant.
}
\description{
\code{Calculate_gold} estimates the derivatives of a dataset according to the Generalized Orthogonal Local Derivative (GOLD)
method described in \url{https://doi.org/10.1080/00273171.2010.498294}
including management of variable time steps.
}
\examples{
time <- c(1:500)/100
signal <- time^2
result <- calculate_gold(signal = signal, time = time, embedding = 5)
}
\keyword{derivative,}
\keyword{embed,}
\keyword{rollmean}
