% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cardio}
\alias{cardio}
\title{Meaurements of cardiac frequency in 21 patients during medical effort tests}
\format{A data frame with 1686 rows of 4 variables
\describe{
  \item{id}{arbitrary identifier of the patient, dimensionless}
  \item{timecol}{time since the beginning of the test, in seconds (s)}
  \item{load}{load of the resistive bycicle, effort that the patient needs to do, in watts (W)}
  \item{hr}{patient's cardiac rythm, in heart beats per minute (1/min)}

}}
\usage{
data(cardio)
}
\description{
A dataset containing time, cardiac frequency and load of the resistive bicycle run by patients during effort tests
}
\keyword{datasets}
