% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{errorcheck}
\alias{errorcheck}
\title{Displays error messages for the analysis function according to the nature of the error
\code{errorcheck} displays error messages and/or warnings concerning the validity of input arguments provided to the analysis function}
\usage{
errorcheck(data, col_var)
}
\arguments{
\item{data}{data.frame or data.table containing the data to be analysed. Same object that is passed as input argument to the analysis function.}

\item{col_var}{column variable. Contains a string that indicates the name of the column to analyze ("id","input",etc.)}
}
\value{
Doesn't return a value. Either displays directly the error message/warning or changes data type in the data.frame/data.table provided
}
\description{
Displays error messages for the analysis function according to the nature of the error
\code{errorcheck} displays error messages and/or warnings concerning the validity of input arguments provided to the analysis function
}
\examples{
input <- "nonexistentname"
errorcheck(cardio,input)
}
