% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{print.doremi}
\alias{print.doremi}
\title{S3 method to print DOREMI objects}
\usage{
\method{print}{doremi}(x, ...)
}
\arguments{
\item{x}{DOREMI object}

\item{...}{includes the additional arguments inherited from the generic print method}
}
\value{
Returns the three coefficients of the differential equation estimated (fixed part, table $resultmean of the DOREMI object)
}
\description{
\code{print.doremi} prints the most important results of a DOREMI object
}
\examples{
myresult <- remi(data = cardio,
                 id = "id",
                 input = "load",
                 time = "time",
                 signal = "hr",
                 embedding = 5)
myresult
}
