% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{print.doremidata}
\alias{print.doremidata}
\title{S3 method to print DOREMIDATA objects}
\usage{
\method{print}{doremidata}(x, ...)
}
\arguments{
\item{x}{DOREMIDATA object}

\item{...}{includes the additional arguments inherited from the generic print method}
}
\value{
Returns the DOREMIDATA object (datatable))
}
\description{
\code{print.doremidata} prints the a DOREMIDATA object
}
\examples{
time <- 0:100
data <- generate.panel.2order(time = time,
                              y0 = 10,
                              v0 = 0,
                              xi = 0.1,
                              period = 30,
                              k = 1,
                              yeq = 2,
                              nind = 6,
                              internoise = 0.3,
                              intranoise = 5)
data
}
