% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremi.R
\name{simulation_generate_order1}
\alias{simulation_generate_order1}
\title{Simulation of various individual signals with intra and inter noise}
\usage{
simulation_generate_order1(nindividuals = 1, dampingtime, amplitude = 1,
  nexc = 1, duration = 10, deltatf = 0.5, tmax, minspacing = 10,
  internoise = 0, intranoise = 0)
}
\arguments{
\item{nindividuals}{Is a positive integer indicating the number of signals that will be simulated, each of them corresponding to an individual.}

\item{dampingtime}{Signal damping time. It is the characteristic response time of the solution to equation (1), and corresponds to the time needed to reach
37\% of the maximum value when there is no excitation (or 63\% of the maximum value for a constant excitation). It should be a signed integer (dampingtime>0).}

\item{amplitude}{is a signed integer or vector of signed integers indicating the amplitude of the pulse. It must be greater than 0,
otherwise an error is generated. It should also have as many elements as the number of pulses (nexc). If this is not the case and the
elements are less than the number of pulses, the amplitude vector will be "recycled" and the elements from it will be repeated until
all the pulses are covered (for instance, if the signal has 6 pulses and the amplitude vector has two elements, pulses 1,3 and 5 will
have the same amplitude as the first element of the amplitude vector and pulses 2,4 and 6 that of the second element).A warning will also
be generated.}

\item{nexc}{is a signed integer indicating the number of pulses to be generated. It must be greater than 0, otherwise an error is generated.}

\item{duration}{is a signed integer or vector of signed integers indicating the duration of the pulse in time units. It must be greater than 0,
otherwise an error is generated. As for the amplitude, it should also have as many elements as the number of pulses (nexc). If this is not the
case and the elements are less than the number of pulses, the amplitude vector will be "recycled" and the elements from it will be repeated until
all the pulses are covered. A warning will also be generated.}

\item{deltatf}{is a signed decimal indicating the time step.}

\item{tmax}{is a signed integer indicating the maximum time range of the excitation in time units. The time vector generated will go from 0 to tmax.}

\item{minspacing}{as pulses are generated randomnly, minspacing is a signed integer that indicates minimum spacing between pulses, thus avoiding
overlapping of the pulses in time. It can be 0 indicating that pulses can follow one another.}

\item{internoise}{Is the inter-individual noise added. The dampingtime accross individuals follows a normal distribution centered on the input parameter dampingtime
with a standard deviation of internoise*dampingtime}

\item{intranoise}{It is the noise added in each individual signal. It also follows a normal distribution with a standard deviation equal to this parameter times the maximum
amplitude of the convolution when using an excitation containing a single pulse (normalisation).}
}
\value{
Returns a list of two objects:
rawdata- contains the function and time values evaluated in nf*precision points. This is to build a pseudo-continuous function with the parameters specified.
data- containts a sampling of raw data at equal time steps to keep only the number of points set in nf.
Each one of these elements contains the following columns:
id - individual identifier (starting by 1 and increasing by 1).
excitation - excitation signal generate through the excitation_function
timecol - time values
dampedsignalraw - signal with no noise (inter noise added for each individual)
dampedsignal - signal with intra noise added
}
\description{
\code{simulation_generate_order1} This function generates signals with intra and inter noise for several individuals.
In order to do this, the function generates a pseudo-continuous signal per individual that is a solution to the first order differential equation
\deqn{\frac{dy(t)}{dt} - \gamma y(t) = E(t)}.
The analytical solution to this equation is a convolution between the Green function and the excitation (see function \code{\link{doremi_generate_order1}})
The function generates a pseudo-continuous signal to increase the precision with which the convolution is calculated. The number of points of this signal
are calculated with the inputs "precision" and "nf". From this expanded signal, the function samples points with a constant time step in order to have
the number of points indicated in nf. These operations are repeated as many times as the value set in the input "nindividuals". Once the signal is sampled,
intra and internoise with normal distributions are added.
}
\details{
Used for simulations in the context of the package.
}
\examples{
simulation_generate_order1(nindividuals = 5,
                           dampingtime = 10,
                           amplitude = c(5,10),
                           nexc = 2,
                           duration = 20,
                           deltatf = 0.5,
                           tmax = 200,
                           minspacing = 0,
                           internoise = 0.2,
                           intranoise = 0.1)
}
\seealso{
\code{\link{doremi_generate_order1}} for calculation of the analytical solution to the differential equation
and \code{\link{excitation_function}} for excitation signal generation
}
\keyword{differential}
\keyword{equation}
\keyword{first}
\keyword{order,}
\keyword{simulation,}
