% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{summary.doremi}
\alias{summary.doremi}
\title{S3 method for DOREMI object summary}
\usage{
\method{summary}{doremi}(object, ...)
}
\arguments{
\item{object}{DOREMI object (contains several lists)}

\item{...}{includes the additional arguments inherited from the generic summary method}
}
\value{
Returns a summary with all the lists of the DOREMI object
}
\description{
\code{summary.doremi} S3 method for the summary function so that it can represent DOREMI objects
}
\examples{
mydata <- simulate.remi(nind = 5,
                           dampingtime = 10,
                           amplitude = c(5,10),
                           nexc = 2,
                           duration = 20,
                           deltatf = 2,
                           tmax = 200,
                           minspacing = 0,
                           internoise = 0.2,
                           intranoise = 0.1)
myresult <- remi(data = mydata,
                           id = "id",
                           input = "excitation",
                           time = "timecol",
                           signal = "dampedsignal",
                           embedding = 5)
summary(myresult)
}
