% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremi.R
\name{analyze.1order}
\alias{analyze.1order}
\title{DOREMI first order analysis function}
\usage{
analyze.1order(
  data,
  id = NULL,
  input = NULL,
  time = NULL,
  signal,
  dermethod = "fda",
  derparam = 2,
  order = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{Is a data frame containing at least one column, that is the signal to be analyzed.}

\item{id}{Is a STRING containing the NAME of the column of data containing the identifier of the individual.
If this parameter is not entered when calling the function, a single individual is assumed and a linear regression is done instead
of the linear mixed-effects regression.}

\item{input}{Is a STRING or a VECTOR OF STRINGS containing the NAME(s) of data column(s) containing the EXCITATION vector(s).
If this parameter is not entered when calling the function,
the excitation is assumed to be unknown. In this case, the linear mixed-effect regression is still carried out but no coefficient is calculated
for the excitation term. The function then uses the parameters estimated by the regression to carry out an exponential fit of the signal
and build the estimated curve.
The function will consider as an excitation each column of data having a name contained in the input vector.
The function will return a coefficient for each one of the excitation variables included in the input vector.}

\item{time}{Is a STRING containing the NAME of the column of data containing the time vector. If this parameter is not entered when calling the function,
it is assumed that time steps are of 1 unit and the time vector is generated internally in the function.}

\item{signal}{Is a STRING containing the NAME of the column of the data frame containing the SIGNAL to be studied.}

\item{dermethod}{is the derivative estimation method. The following methods are available: "gold","glla" and "fda"}

\item{derparam}{If dermethod "glla" or "gold" are chosen, it is the embedding number, a positive integer containing the number of points to be used for the calculation of the derivatives.
Its value by default is 2 as at least two points are needed for the calculation of the first derivative. If dermethod "fda" is chosen, this parameter is
spar, the parameter related to the smoothing parameter lambda used in the penalization function of the function \code{smooth.spline} to estimate the derivative via splines (Functional Data Analysis)}

\item{order}{is the maximum order of the derivative to estimate. Using an order higher than that of the maximum derivative to estimate (1 in first order differential equations and
2 in second order differential equations), for instance, order=4 might enhance derivative estimation (see \href{https://doi.org/10.1080/00273171.2015.1123138}{Chow et al.(2016)})}

\item{verbose}{Is a boolean that displays status messages of the function when set to 1. Useful for debugging.}
}
\value{
Returns a summary of the fixed components estimated by the linear regression for the three coefficients: damping time, excitation coefficient and equilibrium value and the R2 resulting from this estimation
}
\description{
\code{analyze.1order}  estimates the coefficients of a first order differential equation of the form:
\deqn{\frac{dy(t)}{dt} - \gamma (y(t) - yeq) = k*u(t)}
using linear mixed-effect models.
Where y(t) is the individual's signal, \eqn{\dot{y}(t)} is the derivative and u(t) is an external excitation perturbing the dynamics.
}
\details{
The analysis performs the following linear mixed-effects regression:
\deqn{y_{ij}'  \sim   b_{0} +b_{0j}+b_{1} y_{ij}+b_{2} U_{ij}+u_{1j} y_{ij}+u_{2j} U_{ij}+e_{ij}}
with i accounting for the time and j for the different individuals. \eqn{e_{ij}} are the residuals,
\eqn{y_{ij}'} is the derivative calculated on embedding points and
y and U are the signal and the excitation averaged on embedding points.
The coefficients estimated to characterize the signal are calculated as follows:
\itemize{
  \item Damping time, tau:  \eqn{\tau _{j} =  \frac{1}{ \gamma _{j} }}  with \eqn{\gamma _{j} =  b_{1} + u_{1j} }
  \item Gain, k: \eqn{\epsilon _{j} = \frac{b_{2} + u_{2j}}{\gamma _{j}}}. It is the proportionality between the excitation and the
  difference between the maximum value reached by the signal and its initial value.
  \item Equilibrium value, yeq: \eqn{yeq _{j} = \frac{b_{0} + b_{0j}}{\gamma _{j}}}. It is the stable value reached in the absence of excitation.
}
The estimation is performed using the function lmer if there are several individuals or lm if there is only one.
With the above estimated parameters, the estimated signal can be reconstructed for
each individual by using the generate.1order function of this package (based on deSolve's ode function).
The function returns five objects:
\enumerate{
 \item data- A data.frame including the input data, the intermediate calculations used to prepare the variables for
  the fit and the estimated trajectories for each individual.

    signal_rollmean - calculation of the moving average of the signal over embedding points.

    signal_derivate1 - calculation of the first derivative of the signal with the gold,glla or fda methods in embedding points.

    time_derivate - calculation of the moving average of the time vector over embedding points.

    input_rollmean - calculation of the moving average of the excitation vector over embedding points.

    estimated- the estimated signal calculated using deSolve's ode function with a first order model, the excitation provided as input and the damping time,
    excitation coefficient and equilibrium value obtained from the fit.
 \item resultid- A data.frame including for each individual, listed by id number, the damping time, the excitation coefficient and the
 equilibrium value (see variables presented in the Details section).
 \item resultmean- A data.frame including the fixed effects of the three coefficients mentioned above and the R2 resulting from an estimation based on this coefficients
 \item regression- A list containing the summary of the linear mixed-effects regression.

 As seen in the Description section, the print method by default prints only the resultmean element. Each one of the other objects
 can be accessed by indicating $ and their name after the result, for instance, for a DOREMI object called "result", it is possible
 to access the regression summary by typing result$regression.
 \item embedding - contains the embedding number used to generate the results (same as function input argument)
 \item spar - contains the smoothing parameter used for the estimation of the derivatives using splines (method "fda")
}
}
\examples{
myresult <- analyze.1order(data = cardio,
                  id = "id",
                  input = "load",
                  time = "time",
                  signal = "hr",
                  dermethod ="gold",
                  derparam = 5)
}
\seealso{
\code{\link{calculate.gold}\link{calculate.glla}\link{calculate.fda}} to compute the derivatives, for details on embedding/spar.
}
\keyword{analysis,}
\keyword{exponential}
\keyword{first}
\keyword{order,}
