% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremi.R
\name{calculate_gold}
\alias{calculate_gold}
\title{Calculation of derivatives through using the Gold method}
\usage{
calculate_gold(signal, time, embedding = 2)
}
\arguments{
\item{signal}{is a vector containing the values of the function to be derivated.}

\item{time}{Is a vector containing time values.  It has the same length as signal.}

\item{embedding}{is a signed integer indicating the number of points to consider for derivative calculation. Embedding must be greater than 2 because at least
two points are needed for the calculation of the first derivative and at least 3 for the calulation of the second derivative. If it is not the case, the
function displays an error message. If not specified, embedding=2 by default.}
}
\value{
Returns a list containing three columns:
dtime- contains the time values in which the derivative was calculated. That is, the roll mean of the input "time" vector in "embedding" points.
dsignal- contains the calculated derivative of the "signal" vector according to the GOLD method, considering "embedding" points.
embedding- contains the number of points used for the derivative calculation and time roll means, which is constant.
}
\description{
\code{Calculate_gold} calculates the derivatives of a function according to the Gold method described in \url{https://doi.org/10.1080/00273171.2010.498294}
including management of variable time steps.
}
\examples{
time <- c(1:500)/100
signal <- time^2
result <- calculate_gold(signal = signal, time = time, embedding = 5)
}
\keyword{derivative,}
\keyword{embed,}
\keyword{rollmean}
