% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremi.R
\name{doremi_generate_order1}
\alias{doremi_generate_order1}
\title{Generation of the first order differential equation solution}
\usage{
doremi_generate_order1(dampingtime, inputvec, inputtim)
}
\arguments{
\item{dampingtime}{Signal damping coefficient. It represents the characteristic response time of the solution of the differential equation.
It should be a signed integer (dampingtime>0), otherwise a warning will be generated.}

\item{inputvec}{Is a vector containing the values of the excitation signal.}

\item{inputtim}{Is a vector containing the time values of the excitation signal.}
}
\value{
Returns a list containing two elements:
y is a vector containing the values calculated from the convolution of the Green function and the excitation vector.
t is a vector containing the time values
}
\description{
\code{doremi_generate_order1} returns a vector containing the convolution of the Green equation and the excitation function.
}
\examples{
doremi_generate_order1(10,rep(c(0,1,0),c(20,30,50)),1:100)

\dontrun{
doremi_generate_order1(20,matrix(1,2,3),1:100)
}
}
\keyword{Green,}
\keyword{differential}
\keyword{equation}
\keyword{excitation,}
\keyword{first}
\keyword{order}
