% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremi.R
\name{excitation_function}
\alias{excitation_function}
\title{Excitation signal generation}
\usage{
excitation_function(amplitude = 1, nexc = 1, duration = 10, deltat = 1,
  tmax = 100, minspacing = 10)
}
\arguments{
\item{amplitude}{is a signed integer or vector of signed integers indicating the amplitude of the pulse. It must be greater than 0,
otherwise an error is generated. It should also have as many elements as the number of pulses (nexc). If this is not the case and the
elements are less than the number of pulses, the amplitude vector will be "recycled" and the elements from it will be repeated until
all the pulses are covered (for instance, if the signal has 6 pulses and the amplitude vector has two elements, pulses 1,3 and 5 will
have the same amplitude as the first element of the amplitude vector and pulses 2,4 and 6 that of the second element).A warning will also
be generated.}

\item{nexc}{is a signed integer indicating the number of pulses to be generated. It must be greater than 0, otherwise an error is generated.}

\item{duration}{is a signed integer or vector of signed integers indicating the duration of the pulse in time units. It must be greater than 0,
otherwise an error is generated. As for the amplitude, it should also have as many elements as the number of pulses (nexc). If this is not the
case and the elements are less than the number of pulses, the amplitude vector will be "recycled" and the elements from it will be repeated until
all the pulses are covered. A warning will also be generated.}

\item{tmax}{is a signed integer indicating the maximum time range of the excitation in time units. The time vector generated will go from 0 to tmax.}

\item{minspacing}{as pulses are generated randomnly, minspacing is a signed integer that indicates minimum spacing between pulses, thus avoiding
overlapping of the pulses in time. It can be 0 indicating that pulses can follow one another.}

\item{deltatf}{is a signed decimal indicating the time step.}
}
\value{
Returns two vectors:
E- vector containing the values of the excitation generated.
t- vector containing the values of time generated.
}
\description{
\code{excitation_function} returns a vector containing the values of a generated pulsed excitation signal with the pulses randomnly placed
for a given amplitude, duration and spacing between the pulses.
}
\details{
Used for simulations in the context of the package. Beware that the following condition should apply:
\deqn{tmax >= (duration+minspacing)*nexc}
so that the pulses "fit" in the time lapse defined.
The package "seewave" already contains a function that generates a pulse (see function pulsew). The improvement of this function, regarding pulsew is that
it is possible to generate pulses of different duration and amplitude, and not just unitary (amplitude=1).
}
\examples{
excitation_function(c(1,10,20),3,c(1,2,4),0.5,100,10)
excitation_function(3,6,2,1,200,2)
}
\keyword{excitation,}
\keyword{simulation}
