% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremi.R
\name{generate.1order}
\alias{generate.1order}
\title{Generation of the first order differential equation solution with deSolve}
\usage{
generate.1order(
  time = 0:100,
  excitation = as.numeric(0:100 > 50),
  y0 = 0,
  t0 = 0,
  exc0 = 0,
  tau = 10,
  k = 1,
  yeq = 0
)
}
\arguments{
\item{time}{Is a vector containing the time values corresponding to the excitation signal.}

\item{excitation}{Is a vector containing the values of the excitation signal.}

\item{y0}{Signal initial value y(t=t0)}

\item{t0}{Time for the signal initial value y(t=t0), 0 by default but it can be different from 0 or even be absent from the time vector (as long as it is contained
within the interval defined by it)}

\item{exc0}{is the initial value for the excitation: u(t=t0)}

\item{tau}{Signal damping time. It represents the characteristic response time of the solution of the differential equation.
A negative value will produce divergence from equilibrium.}

\item{k}{Signal gain. It represents the proportionnality between the equilibrium value and the input maximum amplitude. It is thus relevant only
for differential equations including an excitation term.}

\item{yeq}{Signal equilibrium value. Value reached when the excitation term is 0 or constant.}
}
\value{
Returns a list containing two elements:
\itemize{
  \item  y is a vector containing the values calculated with deSolve so that y is a solution to a first order differential equation with the constant
  coefficients provided as input.
  \item  t is a vector containing the corresponding time values
}
}
\description{
\code{generate.1order} returns a data frame containing the time (supplied as input) and a simulated signal generated as a solution to a first order
differential equation whith the coefficients are provided as inputs:
\deqn{\frac{dy(t)}{dt} - \gamma (y(t) - yeq) = k*u(t)}
Where y(t) is the signal, dy(t) its derivative, \eqn{\gamma} is the damping rate, k the gain and yeq the equilibrium value.
u(t) is an external excitation perturbing the dynamics.
The latter is also provided as input and it can be null (then the solution
will be a decreasing exponential). The numerical solution is generated with deSolve.
}
\examples{
generate.1order()
generate.1order(t0 = 2.5,y0 = 2)
generate.1order(time = 0:49, excitation = c(rep(0,10),rep(1,40)))
}
\keyword{coefficients}
\keyword{constant}
\keyword{differential}
\keyword{equation,}
\keyword{first}
\keyword{order}
