% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremi.R
\name{generate.2order}
\alias{generate.2order}
\title{Generation of the second order differential equation solution with deSolve}
\usage{
generate.2order(
  time = 0:100,
  excitation = as.numeric(0:100 > 50),
  y0 = 0,
  v0 = 0,
  t0 = 0,
  exc0 = 0,
  xi = 0.1,
  period = 10,
  k = 1,
  yeq = 0
)
}
\arguments{
\item{time}{is a vector containing the time values corresponding to the excitation signal.}

\item{excitation}{Is a vector containing the values of the excitation signal.}

\item{y0}{is the initial condition for the variable y(t=t0), (0, by default), it is a scalar.}

\item{v0}{is the initial condition for the derivative dy(t=t0), (0, by default), it is a scalar.}

\item{t0}{is the initial time}

\item{exc0}{is the initial value for the excitation: u(t=t0)}

\item{xi}{is the damping factor. A negative value will produce divergence from equilibrium.}

\item{period}{is the period T of the oscillation, \eqn{\T = \frac{2\pi}{omega_{n}}} as mentioned}

\item{k}{is the gain. It represents the proportionnality between the equilibrium value and the input maximum amplitude. It is thus relevant only
for differential equations including an excitation term.}

\item{yeq}{is the signal equilibrium value. Value reached when the excitation term is 0 or constant.}
}
\value{
Returns a list containing two elements:
\itemize{
  \item  t is a vector containing the corresponding time values
  \item  y is a vector containing the values calculated with deSolve so that y is a solution to a second order differential equation with constant
  coefficients (provided as input) evaluated at the time points given by t
  \item dy is a vector containing the values of the derivative calculated at the same time points
}
}
\description{
\code{generate.2order} returns a data frame containing the time (supplied as input) and a simulated signal generated as a solution to a second order
differential equation whith constant coefficients that are provided as inputs:
\deqn{\frac{d^2y}{dt} + 2\zeta\omega_{n}\frac{dy}{dt} + \omega_{n}^2 y = k*u(t)}
Where:
y(t) is the signal, dy(t) its derivative and d2y(t) its second derivative
\itemize{
   \item{\eqn{\omega_{n} = \frac{2\pi}{period}} that is the system's natural frequency, the frequency with which the system would vibrate if there were no damping.
   The term \omega_{n}^2 represents thus the ratio between the attraction to the equilibrium and the inertia. If we considered the example
   of a mass attached to a spring, this term would represent the ratio of the spring constant and the object's mass.
   }
   \item{\zeta is the damping ratio. It represents the friction that damps the oscillation of the system (slows the rate of change of the variable).
   The term 2\zeta\omega_n thus represents the respective contribution of the inertia, the friction and the attraction to the equilibrium.
   The value of \zeta determines the shape of the system time response, which can be:
   \zeta<0	Unstable, oscillations of increasing magnitude
   \zeta=0	Undamped, oscillating
   0<\zeta<1	Underdamped or simply "damped". Most of the studies use this model, also referring to it as "Damped Linear Oscillator" (DLO).
   \zeta=1	Critically damped
   \zeta>1	Over-damped, no oscillations in the return to equilibrium
   }
   \item k is the gain
   \item u(t) is an external excitation perturbing the dynamics
}
The excitation is also provided as input and it can be null (then the solution
will be a damped linear oscillator when the initial condition is different from 0)
}
\examples{
generate.2order(time=0:249,excitation=c(rep(0,10),rep(1,240)),period=10)
generate.2order(y0=10)
}
\keyword{coefficients}
\keyword{constant}
\keyword{differential}
\keyword{equation,}
\keyword{order}
\keyword{second}
