% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{plot.doremi}
\alias{plot.doremi}
\title{S3 method to plot DOREMI objects}
\usage{
\method{plot}{doremi}(doremi_object, id = 1:6)
}
\arguments{
\item{doremi_object}{DOREMI object (contains several lists)}

\item{id}{Includes the identifiers of the individuals to be represented in the plot.
id can be a scalar (the function will plot an individual) or a vector (it will plot the individuals with
the numeric id contained in the vector)
By default, it will print the six first individuals.}
}
\value{
Returns a plot with legend and title that includes the name of the DOREMI object
}
\description{
\code{plot.doremi} S3 method for the plot function so that it can represent DOREMI objects
}
\examples{
mydata <- simulation_generate_order1(nindividuals = 5,
                           dampingtime = 10,
                           amplitude = c(5,10),
                           nexc = 2,
                           duration = 20,
                           deltatf = 2,
                           tmax = 200,
                           minspacing = 0,
                           internoise = 0.2,
                           intranoise = 0.1)
myresult <- doremi_analyse_order1(userdata = mydata$data,
                           id = "id",
                           input = "excitation",
                           time = "timecol",
                           signal = "dampedsignal",
                           embedding = 5)
plot(myresult)
plot(myresult,1)
plot(myresult,2:5)
}
