% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{predict.doremi}
\alias{predict.doremi}
\title{S3 method to predict signal values in a DOREMI object when entering a new excitation}
\usage{
\method{predict}{doremi}(object, ..., newdata)
}
\arguments{
\item{object}{DOREMI object result of an analysis with the function doremi_analyse_order1}

\item{...}{Additionnal arguments inherited from generic predict method.}

\item{newdata}{includes a data table containing three columns or more: id (optional), indicating the individual identifier,
time, containing the time values and excitation, being one or several columns containing the different excitations
to which the user desires to obtain an estimated signal. As in the other methods for the predict function, the columns of newdata
must have the same names as those of the original object, otherwise an error message is displayed}
}
\value{
Returns an list containing the values of time, the values of the excitation and the predicted
values of the signal for that excitation
}
\description{
\code{predict.doremi} S3 method for the predict function so that it can
predict signal values in a DOREMI object when entering a new excitation
}
\examples{
mydata <- simulation_generate_order1(nindividuals = 5,
                           dampingtime = 10,
                           amplitude = c(5,10),
                           nexc = 2,
                           duration = 20,
                           deltatf = 2,
                           tmax = 200,
                           minspacing = 0,
                           internoise = 0.2,
                           intranoise = 0.1)
myresult <- doremi_analyse_order1(userdata = mydata$data,
                           id = "id",
                           input = "excitation",
                           time = "timecol",
                           signal = "dampedsignal",
                           embedding = 5)
exc_table <- myresult$data
predresult <- predict(myresult, newdata = exc_table)
plot(predresult)
}
