% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{print.doremi}
\alias{print.doremi}
\title{S3 method to print DOREMI objects}
\usage{
\method{print}{doremi}(x, ...)
}
\arguments{
\item{x}{DOREMI object}

\item{...}{includes the additional arguments inherited from the generic print method}
}
\value{
Returns the coefficients of the differential equation estimated (fixed coefficients, table $resultmean of the DOREMI object)
}
\description{
\code{print.doremi} prints the most important results of a DOREMI object
}
\examples{
myresult <- analyze.1order(data = cardio,
                 id = "id",
                 input = "load",
                 time = "time",
                 signal = "hr")
myresult
}
