% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{print.doremi}
\alias{print.doremi}
\title{S3 method to print DOREMI objects}
\usage{
\method{print}{doremi}(x, ...)
}
\arguments{
\item{x}{DOREMI object (contains several lists)}

\item{...}{includes the additional arguments inherited from the generic print method}
}
\value{
Returns the main parameters of the DOREMI object: the mean values of the three coefficients of the differential equation
}
\description{
\code{print.doremi} S3 method for the print function so that it can represent DOREMI objects
}
\examples{
mydata <- simulate.remi(nind = 5,
                           dampingtime = 10,
                           amplitude = c(5,10),
                           nexc = 2,
                           duration = 20,
                           deltatf = 2,
                           tmax = 200,
                           minspacing = 0,
                           internoise = 0.2,
                           intranoise = 0.1)
myresult <- remi(data = mydata,
                           id = "id",
                           input = "excitation",
                           time = "timecol",
                           signal = "dampedsignal",
                           embedding = 5)
myresult
}
