% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rotation}
\alias{rotation}
\title{Measurements of response time of 17 individuals when carrying out mental rotation tasks}
\format{A data frame with 619 rows and 5 variables
\describe{
  \item{id}{positive integer, arbitrary identifier of the individual}
  \item{sex}{character, sex of the individual, as the study highlighted the difference in response time according to sex}
  \item{numtrials}{positive integer, number of trial to which the mean response time belongs}
  \item{meanRT}{positive integer, mean response time of the individual to execute the mental rotation task, in milliseconds (ms)}
  \item{logmeanRT}{natural logarithm of the mean response time}

}}
\source{
Delphine S. Courvoisier, Olivier Renaud, Christian Geiser, Kerstin Paschke, Kevin Gaudy, Kirsten Jordan,
Sex hormones and mental rotation: An intensive longitudinal investigation,

Hormones and Behavior,

Volume 63, Issue 2,

2013,

Pages 345-351,

https://doi.org/10.1016/j.yhbeh.2012.12.007
}
\usage{
data(rotation)
}
\description{
Data containing reaction time to a mental rotation task over a 60 day period for 17 individuals
\href{https://doi.org/10.1016/j.yhbeh.2012.12.007}{(Courvoisier et al., 2013)}.
}
\keyword{datasets}
