% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{summary.doremi}
\alias{summary.doremi}
\title{S3 method for DOREMI object summary}
\usage{
\method{summary}{doremi}(object, ...)
}
\arguments{
\item{object}{DOREMI object (contains several lists)}

\item{...}{includes the additional arguments inherited from the generic summary method}
}
\value{
Returns a summary containing the five lists of the DOREMI object
}
\description{
\code{summary.doremi} provides a summary of the remi analysis
}
\examples{
myresult <- remi(data = cardio,
                 id = "id",
                 input = "load",
                 time = "time",
                 signal = "hr",
                 embedding = 5)
summary(myresult)
}
